Move EVM runtime components to a module
